Split out examples and benchmarks routes, make scene loading logic a hook
